moved gpu custom operator tests to different file
