+++ tests/python/unittest/test_extensions.py
@@ -166,47 +166,3 @@ def test_subgraph():
     out4 = sym_block(mx.nd.ones((3,2)),mx.nd.ones((3,2)))
     # check that result matches one executed by MXNet
     assert_almost_equal(out[0].asnumpy(), out4[0].asnumpy(), rtol=1e-3, atol=1e-3)
-
-@unittest.skipIf(check_platform(), "not all machine types supported")
-@unittest.skipIf(is_cd_run(), "continuous delivery run - ignoring test")
-@unittest.skipIf(default_context().device_type == 'cpu', "ignoring custom_op_gpu test on cpu run")
-def test_custom_op_gpu():
-    # possible places to find library file
-    if (os.name=='posix'):
-        lib = 'libcustomop_gpu_lib.so'
-        if os.path.exists(lib):
-            fname = lib
-        elif os.path.exists('build/'+lib):
-            fname = 'build/'+lib
-        else:
-            raise MXNetError("library %s not found " % lib)
-    elif (os.name=='nt'):
-        lib = 'libcustomop_gpu_lib.dll'
-        if os.path.exists('windows_package\\lib\\'+lib):
-            fname = 'windows_package\\lib\\'+lib
-        else:
-            raise MXNetError("library %s not found " % lib)
-
-    fname = os.path.abspath(fname)
-    # load the library containing gemm custom operators
-    mx.library.load(fname)
-
-    # test symbol custom relu operator in gpu
-    b = mx.nd.array([[-2,-1],[1,2]], ctx=mx.gpu())
-    c = mx.sym.Variable('c')
-    d = mx.sym.Variable('d')
-    e = mx.sym.my_relu(c)
-    base = mx.sym.relu(d)
-    in_grad = [mx.nd.empty((2,2), ctx=mx.gpu())]
-    in_grad_base = [mx.nd.empty((2,2), ctx=mx.gpu())]
-    exe = e.bind(ctx=mx.gpu(), args={'c':b}, args_grad=in_grad)
-    exe_base = base.bind(ctx=mx.gpu(), args={'d':b}, args_grad=in_grad_base)
-    out = exe.forward()
-    out_base = exe_base.forward()
-    assert_almost_equal(out_base[0].asnumpy(), out[0].asnumpy(), rtol=1e-3, atol=1e-3)
-
-    # test backward
-    out_grad = mx.nd.ones((2,2), ctx=mx.gpu())
-    exe.backward([out_grad])
-    exe_base.backward([out_grad])
-    assert_almost_equal(in_grad_base[0].asnumpy(), in_grad[0].asnumpy(), rtol=1e-3, atol=1e-3)
